Cache products daily in Redis
