+++ .gitignore
@@ -25,3 +25,4 @@ Thumbs.db
 
 # Next.js
 .next/
+next-env.d.ts
