+++ package.json
@@ -7,7 +7,8 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "export": "next export"
+    "export": "next export",
+    "sync:products": "node scripts/sync-products.mjs"
   },
   "dependencies": {
     "@stripe/react-stripe-js": "^2.4.0",
